==> code/00.get_adjacent_gu.py <==
import pickle

adjacent_gu = {'강서구':['강서구','양천구','영등포구'],
               '양천구':['강서구','양천구','영등포구','구로구'],
               '구로구':['양천구', '구로구','영등포구','금천구'],
               '영등포구':['강서구','양천구','영등포구','구로구','동작구'],
               '동작구':['영등포구','동작구','금천구','관악구','서초구'],
               '금천구':['구로구','관악구','금천구'],
               '관악구':['동작구','구로구','금천구','관악구','서초구'],
               '서초구':['동작구','관악구','서초구','강남구'],
               '강남구':['서초구','강남구','송파구'],
               '송파구':['강남구','송파구','강동구'],
               '강동구':['송파구','강동구'],
               '마포구':['마포구','은평구','서대문구','중구','용산구'],
               '용산구':['마포구','중구','용산구','성동구'],
               '성동구':['용산구','중구','성동구','동대문구','광진구'],
               '광진구':['성동구','동대문구','광진구','중랑구'],
               '은평구':['은평구','마포구','서대문구','종로구'],
               '서대문구':['은평구','마포구','서대문구','종로구','중구'],
               '중구':['서대문구','종로구','중구','용산구','동대문구','성동구'],
               '종로구':['은평구','서대문구','종로구','중구','성북구','동대문구'],
               '동대문구':['종로구','성북구','동대문구','성동구','중랑구','광진구'],
               '중랑구':['광진구','동대문구','중랑구','성북구','노원구'],
               '성북구':['종로구','동대문구','성북구','중랑구','강북구','노원구'],
               '강북구':['성북구','강북구','도봉구','노원구'],
               '도봉구':['강북구','노원구','도봉구'],
               '노원구':['도봉구','강북구','노원구','중랑구','성북구']}

with open('../11.data/00.adjacent_gu.pickle','wb') as w:
    pickle.dump(adjacent_gu, w)
